api/requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
pytest==7.4.0
psycopg2-binary==2.9.7
